backend/requirements.txt
fastapi>=0.110.1
uvicorn>=0.29.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
openai>=1.16.2
pydantic>=2.9.0
pydantic-settings>=2.4.0
redis>=5.0.3
